pyproject.toml
[tool.poetry]
name = "mlc-llm"
version = "0.1.0"
description = ""
authors = ["Bella Nicholson <bellanich.software@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
